requirements.txt
absl-py==0.9.0
astor==0.8.1
bleach==1.5.0
certifi==2019.11.28
chardet==3.0.4
git+https://github.com/tensorflow/cleverhans.git#egg=cleverhans
cloudpickle==1.3.0
cox==0.1.post2
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
gast==0.2.2
gitdb==4.0.2
GitPython==3.1.0
google-pasta==0.1.8
GPUtil==1.4.0
grpcio==1.27.2
h5py==2.10.0
html5lib==0.9999999
idna==2.9
imageio==2.8.0
JIT==0.0.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lightfm==1.16
llvmlite==0.31.0
Markdown==3.2.1
matplotlib==3.0.3
mnist==0.2.2
moves==0.1
networkx==2.4
nose==1.3.7
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
opt-einsum==3.1.0
pandas==1.0.1
Pillow==6.2.2
protobuf==3.11.3
psutil==5.8.0
py3nvml==0.2.5
pycodestyle==2.5.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
robustness==1.1.post2
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
smmap==3.0.1
tables==3.6.1
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-probability==0.9.0
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.45.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.0
xmltodict==0.12.0
